[TESTS] Separate standard op tests from test_core.py (#2991)
